+++ towerlib/entities/organization.py
@@ -66,7 +66,7 @@ LOGGER = logging.getLogger(LOGGER_BASENAME)
 LOGGER.addHandler(logging.NullHandler())
 
 
-class Organization(Entity):  # pylint: disable=too-many-public-methods
+class Organization(Entity):
     """Models the organization entity of ansible tower."""
 
     DEFAULT_MEMBER_ROLE = 'Member'
@@ -272,8 +272,7 @@ class Organization(Entity):  # pylint: disable=too-many-public-methods
             project (Project): project on success else None.
 
         """
-        return next(self._tower.projects.filter({'organization': self.id, 'name__iexact': name}), None)\
-
+        return next(self._tower.projects.filter({'organization': self.id, 'name__iexact': name}), None)
 
     def create_project(self,  # pylint: disable=too-many-arguments, too-many-locals
                        name,
@@ -312,7 +311,7 @@ class Organization(Entity):  # pylint: disable=too-many-public-methods
 
         """
         # Credential Type 2 = SCM
-        url = '{api}/projects/'.format(api=self._tower.api)
+        url = f'{self._tower.api}/projects/'
         payload = {'name': name,
                    'description': description,
                    'scm_type': scm_type,